4x speedup to five_point. The new code will give slightly different results to the old code, due to differences in numerics, tolerances and etc.
The nullspace is now found using Gauss-Jordan, instead of SVD (for a speedup
of about 30x for this particular part).

New hand written polynomial root finder, for a speedup of 4x and improved
reliability (the old one would occasionally expect -1 roots due to numerical
instabilities and fail with an alloc_error).

The new root-finder will not attempt to find roots of p(x) for |x|>100, since
it is not possible to find these root to high enough accuracy.

Roots are bracketed using Sturm sequences (as before) but using the resursive
implementation which is more efficient.

The root polishing uses a hybrid Newton-Raphson/bisection rootfinder which
usually gets roots to higher accuracy than the old bisection solver and almost
always converges in < 8 iterations, and often in 4. It will give up after 10
iterations.

A new tolerance feps is introduced, so that for a numerical root at x, if
|p(x)| > feps, the root will not be recorded.

The other tolerances (eps and zeps) follow the convention of NR in C.